requirements.txt
flask==3.0.2
pandas==2.2.1
plotly==5.19.0
matplotlib==3.8.3
numpy==1.26.4
scipy==1.12.0
openpyxl==3.1.2
python-docx==1.1.0
pystray==0.19.5
Pillow==10.2.0
